Refactor entry deletion to pass through diag thread
Fix https://github.com/EFForg/rayhunter/issues/269

Our current deletion system is subject to race conditions. When you
delete an entry, we:

1. send a message to the diag device thread, asynchronously by
   "enqueuing" it into a channel
2. delete the current entry in qmdl store, synchronously

The problem with this is that the diag device thread also accesses the
qmdl store, and may not have seen the StopRecording message yet, even
after step 2.

This can lead to crashes where the diag device thread tries to access
the current entry and fails.

We can add more synchronization primitives, but given the current design
I guess it is best if the diag device thread "wraps" QMDL store for any
write operations, so that it can synchronize internally.

We already did this for StartRecording and StopRecording anyway, so
while it's maybe not ideal that the diag device thread now has to know
about every kind of user action on recordings, we're just making it more
consistent with this.